+++ odpc.client/src/features/publicatie/components/FileUpload.vue
@@ -1,0 +1,11 @@
+<template>
+  <div class="form-group">
+    <label for="bestand">Bestand toevoegen</label>
+
+    <input id="bestand" type="file" class="small" />
+  </div>
+</template>
+
+<script lang="ts" setup></script>
+
+<style lang="scss" scoped></style>
