--- odpc.client/src/views/beheer/GebruikersGroepen.vue
@@ -1,11 +1,0 @@
-<template>
-  <h1>Gebruikersgroepen</h1>
-
-  <gebruikers-groepen />
-</template>
-
-<script setup lang="ts">
-import GebruikersGroepen from "@/features/gebruikersgroep/GebruikersGroepen.vue";
-</script>
-
-<style lang="scss" scoped></style>
