+++ odpc.client/src/views/beheer/GebruikersgroepView.vue
@@ -1,11 +1,11 @@
 <template>
   <h1>Gebruikersgroep</h1>
 
-  <gebruikers-groep :id="id" />
+  <gebruikersgroep-details :id="id" />
 </template>
 
 <script setup lang="ts">
-import GebruikersGroep from "@/features/gebruikersgroep/GebruikersGroep.vue";
+import GebruikersgroepDetails from "@/features/gebruikersgroep/GebruikersgroepDetails.vue";
 
 defineProps<{ id: string }>();
 </script>
