+++ odpc.client/src/views/beheer/GebruikersgroepenView.vue
@@ -1,0 +1,11 @@
+<template>
+  <h1>Gebruikersgroepen</h1>
+
+  <gebruikersgroepen-overview />
+</template>
+
+<script setup lang="ts">
+import GebruikersgroepenOverview from "@/features/gebruikersgroep/GebruikersgroepenOverview.vue";
+</script>
+
+<style lang="scss" scoped></style>
